+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jni/JNIAccessFeature.java
@@ -717,7 +717,7 @@ public class JNIAccessFeature implements Feature {
                     assert hField.equals(hybridLayout.getArrayField()) : "JNI access to hybrid objects is implemented only for the array field";
                     offset = hybridLayout.getArrayBaseOffset();
                 } else {
-                    assert hField.hasLocation();
+                    assert hField.hasLocation() : hField;
                     offset = hField.getLocation();
                 }
                 hidingSubclasses = findHidingSubclasses(hField.getDeclaringClass(), sub -> anyFieldMatches(sub, name));